requirements.txt
Flask==3.1.0
requests==2.32.3
Scrapy==2.12.0
APScheduler==3.10.4
mysql-connector-python==9.2.0
python-dotenv==1.0.1
